Update README and rename packages.
